+++ guillotina_elasticsearch/tests/test_custom_index.py
@@ -11,6 +11,9 @@ import json
 import pytest
 
 
+pytestmark = [pytest.mark.asyncio]
+
+
 async def test_create_index(es_requester):
     async with es_requester as requester:
         resp, status = await requester(
@@ -164,6 +167,7 @@ async def test_delete_resource(es_requester):
 
 async def test_delete_base_removes_index_from_elastic(es_requester):
     async with es_requester as requester:
+        container, request, txn, tm = await setup_txn_on_container(requester)
         cresp, _ = await requester(
             'POST',
             '/db/guillotina/',
@@ -210,7 +214,6 @@ async def test_delete_base_removes_index_from_elastic(es_requester):
             ))
 
 
-@pytest.mark.flaky(reruns=5)
 async def test_delete_parent_gets_cleaned_up(es_requester):
     async with es_requester as requester:
         await requester(