+++ guillotina_elasticsearch/tests/test_search.py
@@ -2,6 +2,9 @@ from guillotina_elasticsearch.tests.utils import run_with_retries
 from guillotina_elasticsearch.tests.utils import setup_txn_on_container
 
 import json
+import pytest
+
+pytestmark = [pytest.mark.asyncio]
 
 
 async def test_indexing_and_search(es_requester):
@@ -30,8 +33,9 @@ async def test_indexing_and_search(es_requester):
                 '/db/guillotina/@search',
                 data=json.dumps({})
             )
-            assert resp['items_count'] == 1
-            assert resp['member'][0]['path'] == '/item1'
+            assert status == 200
+            assert resp['items_total'] == 1
+            assert resp['items'][0]['path'] == '/item1'
 
         await run_with_retries(_test, requester)
 
@@ -44,7 +48,7 @@ async def test_indexing_and_search(es_requester):
                 '/db/guillotina/@search',
                 data=json.dumps({})
             )
-            assert resp['items_count'] == 0
+            assert resp['items_total'] == 0
 
         await run_with_retries(_test, requester)
 
@@ -87,8 +91,8 @@ async def test_removes_all_children(es_requester):
                 '/db/guillotina/@search',
                 data=json.dumps({})
             )
-            assert resp['items_count'] == 3
-            assert resp['member'][0]['@name']
+            assert resp['items_total'] == 3
+            assert resp['items'][0]['@name']
 
         await run_with_retries(_test, requester)
 
@@ -101,6 +105,6 @@ async def test_removes_all_children(es_requester):
                 '/db/guillotina/@search',
                 data=json.dumps({})
             )
-            assert resp['items_count'] == 0
+            assert resp['items_total'] == 0
 
         await run_with_retries(_test, requester)