+++ guillotina_elasticsearch/tests/test_vacuum.py
@@ -14,11 +14,13 @@ import os
 import pytest
 
 
+pytestmark = [pytest.mark.asyncio]
+
+
 DATABASE = os.environ.get('DATABASE', 'DUMMY')
 
 
 @pytest.mark.skipif(DATABASE == 'DUMMY', reason='Not for dummy db')
-@pytest.mark.flaky(reruns=5)
 async def test_adds_missing_elasticsearch_entry(es_requester):
     async with es_requester as requester:
         await add_content(requester)
@@ -59,7 +61,6 @@ async def test_adds_missing_elasticsearch_entry(es_requester):
 
 
 @pytest.mark.skipif(DATABASE == 'DUMMY', reason='Not for dummy db')
-@pytest.mark.flaky(reruns=5)
 async def test_updates_out_of_data_es_entries(es_requester):
     async with es_requester as requester:
         await add_content(requester)
@@ -97,7 +98,6 @@ async def test_updates_out_of_data_es_entries(es_requester):
 
 
 @pytest.mark.skipif(DATABASE == 'DUMMY', reason='Not for dummy db')
-@pytest.mark.flaky(reruns=5)
 async def test_removes_orphaned_es_entry(es_requester):
     async with es_requester as requester:
         container, request, txn, tm = await setup_txn_on_container(requester)
@@ -132,7 +132,6 @@ async def test_removes_orphaned_es_entry(es_requester):
 
 
 @pytest.mark.skipif(DATABASE == 'DUMMY', reason='Not for dummy db')
-@pytest.mark.flaky(reruns=5)
 async def test_vacuum_with_sub_indexes(es_requester):
     async with es_requester as requester:
         await add_content(
@@ -219,7 +218,6 @@ async def test_vacuum_with_sub_indexes(es_requester):
 
 
 @pytest.mark.skipif(DATABASE == 'DUMMY', reason='Not for dummy db')
-@pytest.mark.flaky(reruns=3)
 async def test_reindexes_moved_content(es_requester):
     async with es_requester as requester:
         resp1, _ = await requester(
@@ -331,7 +329,6 @@ async def test_reindexes_moved_content(es_requester):
 
 
 @pytest.mark.skipif(DATABASE == 'DUMMY', reason='Not for dummy db')
-# @pytest.mark.flaky(reruns=3)
 async def test_vacuum_with_multiple_containers(es_requester):
     async with es_requester as requester: